Tests/LibWeb/Text/expected/control-message-queue.txt

AudioContext state: suspended
AudioContext.suspend: function
AudioContext.resume: function
AudioContext.close: function
suspend() returns Promise: true
OfflineAudioContext channels: 2
OfflineAudioContext.startRendering: function
